PeopleCounterApp/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
futures==3.1.1
idna==2.9
networkx==2.4
numpy==1.18.3
paho-mqtt==1.5.0
pyspin==1.1.1
PyYAML==5.3.1
requests==2.23.0
urllib3==1.25.9
